Fix power consumption in sleep mode for RFM69.
Apparently before going to sleep in some cases, you must re-set the RegTestPa1 or
RegTestPa2 registers. If you do this, you get the spec'd sleep current
(0.1uA, or as close as my gear can measure). Without doing this, my
measured current draw was ~82uA. My test case was:

- Init radio, set power level (actual power level didn't matter) and
frequency.
- At this point, making the radio sleep has expected behavior - i.e. it
consumes ~0.1uA
- Send a packet
- Put radio to sleep
- Now, radio consume ~82uA

Setting either of RegTestPa1 or RegTestPa2 before going to sleep the
second time was sufficient to fix this. I'm not sure why this works, so
this change sets both.